email_ingest/requirements.txt
aws-embedded-metrics==3.0.0
aws-lambda-logging==0.1.1
boto3==1.26.87
botocore==1.29.33
bs4==0.0.1
certifi==2022.12.7
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
email-reply-parser==0.5.12
html5lib==1.1
humanreadable==0.1.0
icalendar==5.0.3
lxml==4.9.2
mock==4.0.3
python-dotenv==0.21.0
salmon-mail==3.2.0
six==1.16.0
tika==1.24
urllib3==1.26.8
vaderSentiment==3.3.2
